Removed unneeded import of DocumentView from tests. Moved SOLR query methods to base_view. Modified config to include specific variable for BIGQUERY_MAX_ROWS
