+++ biblib/views/library_view.py
@@ -139,73 +139,6 @@ class LibraryView(BaseView):
         return False
 
     @staticmethod
-    def solr_big_query(
-            bibcodes,
-            start=0,
-            rows=20,
-            sort='date desc',
-            fl='bibcode'
-    ):
-        """
-        A thin wrapper for the solr bigquery service.
-
-        :param bibcodes: bibcodes
-        :type bibcodes: list
-
-        :param start: start index
-        :type start: int
-
-        :param rows: number of rows
-        :type rows: int
-
-        :param sort: how the response should be sorted
-        :type sort: str
-
-        :param fl: Solr fields to be returned
-        :type fl: str
-
-        :return: solr bigquery end point response
-        """
-
-        bibcodes_string = 'bibcode\n' + '\n'.join(bibcodes)
-
-        # We need atleast bibcode and alternate bibcode for other methods
-        # to work properly
-        if fl == '':
-            fl = 'bibcode,alternate_bibcode'
-        else:
-            fl_split = fl.split(',')
-            for required_fl in ['bibcode', 'alternate_bibcode']:
-                if required_fl not in fl_split:
-                    fl = '{},{}'.format(fl, required_fl)
-
-        params = {
-            'q': '*:*',
-            'wt': 'json',
-            'fl': fl,
-            'rows': rows,
-            'start': start,
-            'fq': '{!bitset}',
-            'sort': sort
-        }
-
-        headers = {
-            'Content-Type': 'big-query/csv',
-            'Authorization': current_app.config.get('SERVICE_TOKEN', request.headers.get('X-Forwarded-Authorization', request.headers.get('Authorization', '')))
-        }
-        current_app.logger.info('Querying Solr bigquery microservice: {0}, {1}'
-                                .format(params,
-                                        bibcodes_string.replace('\n', ',')))
-        response = client().post(
-            url=current_app.config['BIBLIB_SOLR_BIG_QUERY_URL'],
-            params=params,
-            data=bibcodes_string,
-            headers=headers
-        )
-
-        return response
-
-    @staticmethod
     def solr_update_library(library_id, solr_docs):
         """
         Updates the library based on the solr canonical bibcodes response
